Refactored test auth token generation to util file
